requirements.txt
# Groq API - İçerik üretimi
groq==0.4.2

# HTTP istekleri
aiohttp==3.9.1
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Görsel işleme
Pillow==10.1.0

# Progress bar
tqdm==4.66.1

# Google Indexing API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0

# Rate limiting
ratelimit==2.2.1

# YAML işleme
PyYAML==6.0.1
